refactor: break apart MapSearch file into components
