+++ dev-client/src/screens/SitesScreen/components/search/MapSearchSuggestionBox.tsx
@@ -1,0 +1,53 @@
+/*
+ * Copyright © 2024 Technology Matters
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see https://www.gnu.org/licenses/.
+ */
+
+import {useCallback} from 'react';
+
+import {Pressable} from 'native-base';
+
+import {
+  Column,
+  Text,
+} from 'terraso-mobile-client/components/NativeBaseAdapters';
+
+type Props = {
+  name: string;
+  address: string;
+  mapboxId: string;
+  onPress: (name: string, mapboxId: string) => void;
+};
+
+export const MapSearchSuggestionBox = ({
+  name,
+  address,
+  mapboxId,
+  onPress,
+}: Props) => {
+  const selectSuggestion = useCallback(
+    () => onPress(name, mapboxId),
+    [name, mapboxId, onPress],
+  );
+
+  return (
+    <Pressable w="100%" py={1} px={3} onPress={selectSuggestion}>
+      <Column>
+        <Text>{name}</Text>
+        <Text>{address}</Text>
+      </Column>
+    </Pressable>
+  );
+};
